perf(frontend): isolate sparkles effect to prevent global re-renders
