+++ src/App.tsx
@@ -18,6 +18,7 @@ import Navbar from "./components/Navbar";
 import Chatbot from "./components/Chatbot";
 import StreakBadge from "./components/StreakBadge";
 import FloatingAI from "./components/FloatingAI";
+import Sparkles from "./components/Sparkles";
 
 import { useAuth } from "@/contexts/useAuth";
 
@@ -68,48 +69,10 @@ const WithNav = ({ children }: { children: React.ReactNode }) => {
 
 function AppContent() {
   const { user } = useAuth();
-  const [sparkles, setSparkles] = useState<{ id: number; x: number; y: number }[]>([]);
-  const sparkleIdRef = useRef(0);
-
-  useEffect(() => {
-    let timeouts: NodeJS.Timeout[] = [];
-    
-    const handleMouseMove = (e: MouseEvent) => {
-      const newSparkles = Array.from({ length: 2 }).map(() => ({
-        id: sparkleIdRef.current++,
-        x: e.clientX + Math.random() * 10 - 5,
-        y: e.clientY + Math.random() * 10 - 5,
-      }));
-
-      setSparkles((prev) => [...prev, ...newSparkles]);
-
-      newSparkles.forEach((sparkle) => {
-        const timeout = setTimeout(() => {
-          setSparkles((prev) => prev.filter((s) => s.id !== sparkle.id));
-        }, 800);
-        timeouts.push(timeout);
-      });
-    };
-
-    window.addEventListener("mousemove", handleMouseMove);
-
-    return () => {
-      window.removeEventListener("mousemove", handleMouseMove);
-      timeouts.forEach(clearTimeout);
-    };
-  }, []);
 
   return (
     <>
-      <div id="sparkle-container">
-        {sparkles.map((sparkle) => (
-          <div
-            key={sparkle.id}
-            className="sparkle"
-            style={{ left: `${sparkle.x}px`, top: `${sparkle.y}px`, position: 'absolute', pointerEvents: 'none' }}
-          />
-        ))}
-      </div>
+      <Sparkles />
 
 
       <Suspense fallback={<div className="flex min-h-screen items-center justify-center bg-[#020617]"><div className="h-10 w-10 animate-spin rounded-full border-4 border-cyan-400 border-t-transparent" /></div>}>
